sched-m3: start threads with basic stack frame
On cores with floating point registers, the initial stack frame is an
extended frame.  However, we only construct a basic frame for new
threads.  Because we need to retain the lr for a given thread, we cannot
store lr on the supervisor stack anymore.  Store the lr in the stack
frame as well.